# Add blocksparse_int_addmm. Eliminate unnecessary contiguous calls which leads to performance increase.