--- a/torchao/sparsity/prototype/superblock/evaluate.py
+++ b/torchao/sparsity/prototype/superblock/evaluate.py
@@ -16,7 +16,7 @@ from torchvision.transforms.functional import InterpolationMode
 
 from torchao.sparsity import sparsify_, semi_sparse_weight
 from torchao.sparsity.prototype.superblock.supermask import apply_supermask
-from torchao.sparsity.prototype.superblock.utils import apply_sparsity, verify_sparsity, mlp_only_with_args
+from torchao.sparsity.prototype.superblock.utils import apply_sparsity, verify_sparsity, mlp_only_with_args, simulate_sparsity, accelerate_with_sparsity
 from torchao.sparsity.prototype.superblock.train import evaluate, _get_cache_path, load_data
 from torchao.sparsity.prototype.sparsifier.weight_norm_sparsifier import WeightNormSparsifier
 
@@ -35,7 +35,7 @@ def main(args):
     val_dir = os.path.join(args.data_path, "val")
     dataset_test, test_sampler = load_data(None, val_dir, args)
     data_loader_test = torch.utils.data.DataLoader(
-        dataset_test, batch_size=args.batch_size, sampler=test_sampler, num_workers=args.workers, pin_memory=True
+        dataset_test, batch_size=args.batch_size, sampler=test_sampler, num_workers=args.workers, pin_memory=True, drop_last=True,
     )
     num_classes = len(dataset_test.classes)
 
@@ -56,7 +56,7 @@ def main(args):
     model.to(device).bfloat16()
 
     if sparsifier_or_none is not None:
-        sparsifier.squash_mask()
+        sparsifier_or_none.squash_mask()
     accelerate_with_sparsity(model, args)
             
     criterion = nn.CrossEntropyLoss(label_smoothing=args.label_smoothing)
@@ -110,6 +110,7 @@ def get_args_parser(add_help=True):
     parser.add_argument("--skip-first-transformer-sparsity", action="store_true", help="Skip applying sparsity to the first transformer layer (for vit only)")
     parser.add_argument('--bsr', type=int, nargs='?', default=64, help='Convert sparsified weights to BSR format with optional block size (default: 64)')
     parser.add_argument('--meta', action='store_true', help='Use Meta internal imagenet structure')
+    parser.add_argument("--quantization", action="store_true", help="Profile the run and dump Prefetto trace")   
 
     return parser