As in the title.

This PR is created on top of https://github.com/pytorch/ao/pull/821 and requires https://github.com/pytorch/pytorch/pull/136104 .

As a result, the following performance test
```
python torchao/sparsity/prototype/superblock/benchmark.py --model vit_h_14   --batch-size 256   --sparsity-linear 0.8   --sp-linear-tile-size 64 --bsr 64 --sparsity bsr --quantization  --tune-kernel-params
```
leads to
```
340.710 ms
2.935 img/s
Memory: 2592
```
which should be compared to the previous state:
```
581.503 ms
1.720 img/s
Memory: 5004
```